Issue #3 Add SourceLocation annotation
This is a quick adaptation from the approach here
https://stackoverflow.com/a/19613827/85371

It's not as pretty as I'd like. It's probably time to introduce proper
constructors for Ast nodes. Until this time, the AST node could be
thought of as "intuitive 1:1 mappings of the common concepts they
represent" but SourceLocation really puts the aggregate initializer
lists a bit over the top for me.

I have not added anything to exercise the SourceLocation information.
But the population is there. Of course you can easily modify the
SourceLocation struct to contain some other meta data. I just thought
that a string_view made sense due to the choice for `std::string` as the
interface type for text to be parsed.